requirements.txt
opencv-python
pillow
pygame
numpy
